Fix "make test" with SIGTERM ignored.
If the process that exec'd "make test" had SIGTERM set to SIG_IGN, one
test failed.  That not a clearly-valid use case.  Even so, since all but
one test already survived that scenario, make the last test cope.
https://www.cpantesters.org/report/efbcc1aa-f5f1-11ee-a642-ee3ed50263fb
inspired this.  That smoke tester starting failing when it upgraded from
NetBSD 9.3 to NetBSD 10.0.  I don't know the cause of the failures; this
may or may not fix them.